src/main/java/ru/job4j/it/EvenNumbersIterator.java
package ru.job4j.it;

import java.util.Arrays;
import java.util.Iterator;
import java.util.NoSuchElementException;
import java.util.function.Consumer;

public class EvenNumbersIterator implements Iterator<Integer> {

    private int[] data;
    private int index;

    public EvenNumbersIterator(int[] data) {
        this.data = data;
    }

    @Override
    public boolean hasNext() {
        return Arrays.stream(data)
                .skip(index)
                .filter(el -> el % 2 == 0)
                .findAny()
                .isPresent();
    }

    @Override
    public Integer next() throws NoSuchElementException {
        if (!hasNext()) {
            throw new NoSuchElementException();
        }
        int rsl = data[index++];
        return rsl % 2 == 0 ? rsl : next();
    }

    @Override
    public void remove() throws UnsupportedOperationException {
        throw new UnsupportedOperationException();
    }

    @Override
    public void forEachRemaining(Consumer<? super Integer> action) throws UnsupportedOperationException {
        throw new UnsupportedOperationException();
    }

}